LawyerOrganizer with StoryBoard
Just adding the app with the StoryBoard this time, I’m still having an
issue with the Organizer view because I’m not sure of what tools to use
to create a recognizable organizer, though navigating between views is
easy enough.  Everything else still looks about the same other than a
main page and additional buttons to navigate.