Week 5
Added the Archiving and it seems to work fine, I’m a little confused as
to why the need for sql but I’ll probably find out why in the readings.
 Design wise it is the same but I did alter Story Board a little to get
the table view to display the entered information.  I’m trying to
design something like drop down boxes to work with the Organizer
portion of the app.